libswoc: update to 1.4.0
